Attempt to address intermittent failure:
WebSocketServer tests close() when done, but before that happens a query is made to send Games out. When the query completes, SocketNotifier is called. Presumably the websocket *exists* but has not yet been removed from config.sockets. NOTE: This may be a memory leak? Are we maybe not removing sockets when they close? Or is this intermittent because it happens JUST as the websocket is closing?